backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.46.0",
    "sqlmodel>=0.0.38",
    "psycopg2-binary>=2.9.12",
    "bcrypt>=5.0.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "pytest>=8.4.2",
    "httpx>=0.28.1",
    "ruff>=0.15.12",
]
